+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Upcoming
 
 * Support multiple errors within a text
 * Add parameters to #bbcode_check_validity to add/remove tags
+* Support regular tag parameters: [tag param=value][/tag]
 
 Version 1.0.1 - 04-Jan-2015